SYKA/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
WTForms==3.1.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
loguru==0.7.2
python-dotenv==1.0.1
redis==5.0.1
celery==5.3.4
gunicorn==22.0.0
pyotp==2.9.0
bcrypt==4.0.1
cryptography==41.0.7
pysocks==1.7.1
telethon==1.36.0
cryptg==0.4.0
psutil==5.9.8
openpyxl==3.1.2
requests==2.31.0
Flask-Migrate==4.0.5